# services/external/api-gateway/api/server/api.yaml
openapi: 3.0.3
info:
  title: codex-k8s api-gateway
  version: 0.2.0
  description: |
    External webhook + staff API contract for codex-k8s.
servers:
  - url: https://platform.codex-k8s.dev
paths:
  /api/v1/webhooks/github:
    post:
      summary: Ingest GitHub webhook delivery
      operationId: ingestGithubWebhook
      tags: [webhooks]
      parameters:
        - in: header
          name: X-GitHub-Event
          required: true
          schema:
            type: string
        - in: header
          name: X-GitHub-Delivery
          required: true
          schema:
            type: string
        - in: header
          name: X-Hub-Signature-256
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          description: Accepted and enqueued for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestGitHubWebhookResponse"
        "200":
          description: Duplicate delivery already ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestGitHubWebhookResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/mcp/approver/callback:
    post:
      summary: Resolve one MCP approval request from external approver callback
      operationId: mcpApproverCallback
      tags: [mcp-callbacks]
      parameters:
        - $ref: "#/components/parameters/MCPCallbackToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MCPApprovalCallbackRequest"
      responses:
        "200":
          description: Approval callback applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveApprovalDecisionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/mcp/executor/callback:
    post:
      summary: Resolve one MCP approval request from external executor callback
      operationId: mcpExecutorCallback
      tags: [mcp-callbacks]
      parameters:
        - $ref: "#/components/parameters/MCPCallbackToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MCPApprovalCallbackRequest"
      responses:
        "200":
          description: Executor callback applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveApprovalDecisionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/github/login:
    get:
      summary: Start GitHub OAuth login
      operationId: loginGithub
      tags: [auth]
      responses:
        "302":
          description: Redirect to GitHub OAuth authorize URL
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/github/callback:
    get:
      summary: Complete GitHub OAuth callback and set auth cookie
      operationId: callbackGithub
      tags: [auth]
      parameters:
        - in: query
          name: state
          required: true
          schema:
            type: string
        - in: query
          name: code
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to staff UI after successful login
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/logout:
    post:
      summary: Logout current staff user
      operationId: logout
      tags: [auth]
      responses:
        "204":
          description: Logged out
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/me:
    get:
      summary: Get current authenticated staff principal
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Authenticated principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/staff/projects:
    get:
      summary: List projects
      operationId: listProjects
      tags: [staff-projects]
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectItemsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create or update a project by slug
      operationId: upsertProject
      tags: [staff-projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProjectRequest"
      responses:
        "201":
          description: Project created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/staff/projects/{project_id}:
    get:
      summary: Get project by id
      operationId: getProject
      tags: [staff-projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete project by id
      operationId: deleteProject
      tags: [staff-projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "204":
          description: Project deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/runs/{run_id}/bypass:
    get:
      summary: Get run by run access key (OAuth bypass)
      operationId: getRunByAccessKey
      tags: [run-bypass]
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/RunAccessKeyQuery"
        - $ref: "#/components/parameters/BypassNamespaceQuery"
        - $ref: "#/components/parameters/BypassTargetEnvQuery"
        - $ref: "#/components/parameters/BypassRuntimeModeQuery"
      responses:
        "200":
          description: Run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/runs/{run_id}/bypass/events:
    get:
      summary: List run flow events by run access key (OAuth bypass)
      operationId: listRunEventsByAccessKey
      tags: [run-bypass]
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/RunAccessKeyQuery"
        - $ref: "#/components/parameters/BypassNamespaceQuery"
        - $ref: "#/components/parameters/BypassTargetEnvQuery"
        - $ref: "#/components/parameters/BypassRuntimeModeQuery"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Flow events list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowEventItemsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/runs/{run_id}/bypass/logs:
    get:
      summary: Get run logs by run access key (OAuth bypass)
      operationId: getRunLogsByAccessKey
      tags: [run-bypass]
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/RunAccessKeyQuery"
        - $ref: "#/components/parameters/BypassNamespaceQuery"
        - $ref: "#/components/parameters/BypassTargetEnvQuery"
        - $ref: "#/components/parameters/BypassRuntimeModeQuery"
        - $ref: "#/components/parameters/TailLines"
      responses:
        "200":
          description: Run logs snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunLogs"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/staff/runs:
    get:
      summary: List runs
      operationId: listRuns
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Runs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunItemsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/staff/runs/jobs:
    get:
      summary: List running jobs
      operationId: listRunJobs
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/TriggerKindFilter"
        - $ref: "#/components/parameters/RunStatusFilter"
        - $ref: "#/components/parameters/AgentKeyFilter"
      responses:
        "200":
          description: Running jobs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunItemsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/staff/runs/waits:
    get:
      summary: List wait queue runs
      operationId: listRunWaits
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/TriggerKindFilter"
        - $ref: "#/components/parameters/RunStatusFilter"
        - $ref: "#/components/parameters/AgentKeyFilter"
        - $ref: "#/components/parameters/WaitStateFilter"
      responses:
        "200":
          description: Wait queue list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunItemsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/staff/approvals:
    get:
      summary: List pending approval requests
      operationId: listPendingApprovals
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Pending approvals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequestItemsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/staff/approvals/{approval_request_id}/decision:
    post:
      summary: Resolve one approval request (approve/deny/expire/fail)
      operationId: resolveApprovalDecision
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/ApprovalRequestID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveApprovalDecisionRequest"
      responses:
        "200":
          description: Approval resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveApprovalDecisionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/staff/runs/{run_id}:
    get:
      summary: Get run by id
      operationId: getRun
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/staff/runs/{run_id}/access-key:
    get:
      summary: Get run access key status
      operationId: getRunAccessKeyStatus
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Run access key status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunAccessKeyStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/staff/runs/{run_id}/access-key/regenerate:
    post:
      summary: Regenerate run access key
      operationId: regenerateRunAccessKey
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/RunID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateRunAccessKeyRequest"
      responses:
        "200":
          description: Run access key rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunAccessKeyIssueResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/staff/runs/{run_id}/access-key/revoke:
    post:
      summary: Revoke run access key
      operationId: revokeRunAccessKey
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Run access key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunAccessKeyStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/staff/runs/{run_id}/namespace:
    delete:
      summary: Force delete run namespace
      operationId: deleteRunNamespace
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Namespace cleanup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunNamespaceCleanupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/staff/runs/{run_id}/events:
    get:
      summary: List run flow events
      operationId: listRunEvents
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Flow events list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowEventItemsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/staff/runs/{run_id}/logs:
    get:
      summary: Get run logs snapshot with tail lines
      operationId: getRunLogs
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/TailLines"
      responses:
        "200":
          description: Run logs snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunLogs"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/staff/runs/{run_id}/learning-feedback:
    get:
      summary: List run learning feedback
      operationId: listRunLearningFeedback
      tags: [staff-runs]
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Learning feedback list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningFeedbackItemsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/staff/runtime-deploy/tasks:
    get:
      summary: List runtime deploy tasks
      operationId: listRuntimeDeployTasks
      tags: [staff-runtime-deploy]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, running, succeeded, failed]
        - in: query
          name: target_env
          schema:
            type: string
      responses:
        "200":
          description: Runtime deploy tasks list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeDeployTaskItemsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/staff/runtime-deploy/tasks/{run_id}:
    get:
      summary: Get runtime deploy task details
      operationId: getRuntimeDeployTask
      tags: [staff-runtime-deploy]
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Runtime deploy task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeDeployTask"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/staff/runtime-errors:
    get:
      summary: List runtime error journal entries
      operationId: listRuntimeErrors
      tags: [staff-runtime-deploy]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - in: query
          name: state
          schema:
            type: string
            enum: [active, viewed, all]
        - in: query
          name: level
          schema:
            type: string
            enum: [error, warning, critical]
        - in: query
          name: source
          schema:
            type: string
        - in: query
          name: run_id
          schema:
            type: string
        - in: query
          name: correlation_id
          schema:
            type: string
      responses:
        "200":
          description: Runtime error journal entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeErrorItemsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/staff/runtime-errors/{runtime_error_id}/viewed:
    post:
      summary: Mark runtime error as viewed
      operationId: markRuntimeErrorViewed
      tags: [staff-runtime-deploy]
      parameters:
        - $ref: "#/components/parameters/RuntimeErrorID"
      responses:
        "200":
          description: Runtime error marked as viewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeError"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/staff/users/{user_id}:
    delete:
      summary: Delete user
      operationId: deleteUser
      tags: [staff-users]
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "204":
          description: User deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
...
    RunID:
      in: path
      name: run_id
      required: true
      schema:
        type: string
...
    Limit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
...
    MCPCallbackToken:
      in: header
      name: X-Codex-MCP-Token
      required: false
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Internal:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    IngestGitHubWebhookResponse:
      type: object
      additionalProperties: false
      required: [correlation_id, status, duplicate]
      properties:
        correlation_id:
          type: string
        run_id:
          type: string
        status:
          type: string
          enum: [accepted, duplicate, ignored]
        duplicate:
          type: boolean
...
    MeResponse:
      type: object
      additionalProperties: false
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/MeUser"
...
    MCPApprovalCallbackRequest:
      type: object
      additionalProperties: false
      required: [approval_request_id, decision]
      properties:
        approval_request_id:
          type: integer
          format: int64
        decision:
          type: string
          enum: [approved, denied, expired, failed, applied]
        reason:
          type: string
          nullable: true
        actor_id:
          type: string
          nullable: true
...